+++ src/main/java/hu/drumbun/controller/need/model/NeedModel.java
@@ -1,4 +1,4 @@
-package hu.drumbun.controller.need;
+package hu.drumbun.controller.need.model;
 
 import hu.drumbun.entities.Path;
